Improved favorite button.
Updated some verbiage on the Dashboard screen and minor things.

Updated some verbiage on the Home page.

Updated main.py to reflect the backend necessity to save photos.

Updated recipedetail.jsx to show the picture of recipe if uploaded, but if no picture uploaded, the recipe title stays to the left; if there is a picture then recipe title goes central.

Updated the recipeform.jsx to include the option/ability to take a pic or upload a pic to the recipe